Fix state management
